requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.11
# by the following command:
#
#    pip-compile requirements.in
#
aiohttp==3.9.3
    # via
    #   -r requirements.in
    #   renault-api
aiosignal==1.3.1
    # via aiohttp
andersen-ev==0.1.4
    # via -r requirements.in
anyio==4.3.0
    # via gql
arrow==1.3.0
    # via async-cron
async-cron==1.6.2
    # via -r requirements.in
asyncio==3.4.3
    # via -r requirements.in
attrs==23.2.0
    # via aiohttp
backoff==2.2.1
    # via gql
black==24.2.0
    # via -r requirements.in
boto3==1.34.51
    # via pycognito
botocore==1.34.51
    # via
    #   boto3
    #   s3transfer
build==1.0.3
    # via pip-tools
certifi==2024.2.2
    # via requests
cffi==1.16.0
    # via cryptography
charset-normalizer==3.3.2
    # via requests
click==8.1.7
    # via
    #   black
    #   pip-tools
cryptography==42.0.5
    # via pyjwt
dateparser==1.2.0
    # via -r requirements.in
envs==1.4
    # via pycognito
frozenlist==1.4.1
    # via
    #   aiohttp
    #   aiosignal
gql==3.5.0
    # via -r requirements.in
graphql-core==3.2.3
    # via gql
idna==3.6
    # via
    #   anyio
    #   requests
    #   yarl
iniconfig==2.0.0
    # via pytest
jmespath==1.0.1
    # via
    #   boto3
    #   botocore
marshmallow==3.21.0
    # via marshmallow-dataclass
marshmallow-dataclass==8.6.0
    # via renault-api
multidict==6.0.5
    # via
    #   aiohttp
    #   yarl
mypy-extensions==1.0.0
    # via
    #   black
    #   typing-inspect
packaging==23.2
    # via
    #   black
    #   build
    #   marshmallow
    #   pytest
pathspec==0.12.1
    # via black
pip-tools==7.4.0
    # via -r requirements.in
platformdirs==4.2.0
    # via black
pluggy==1.4.0
    # via pytest
pycognito==2024.2.0
    # via -r requirements.in
pycparser==2.21
    # via cffi
pyjwt[crypto]==2.8.0
    # via
    #   pycognito
    #   renault-api
pyproject-hooks==1.0.0
    # via
    #   build
    #   pip-tools
pytest==8.0.2
    # via
    #   -r requirements.in
    #   pytest-asyncio
pytest-asyncio==0.23.5
    # via -r requirements.in
python-dateutil==2.8.2
    # via
    #   arrow
    #   botocore
    #   dateparser
python-dotenv==1.0.1
    # via -r requirements.in
pytz==2024.1
    # via dateparser
regex==2023.12.25
    # via dateparser
renault-api==0.2.1
    # via -r requirements.in
requests==2.31.0
    # via
    #   -r requirements.in
    #   pycognito
    #   requests-toolbelt
requests-toolbelt==1.0.0
    # via -r requirements.in
s3transfer==0.10.0
    # via boto3
six==1.16.0
    # via python-dateutil
sniffio==1.3.1
    # via anyio
tabulate==0.9.0
    # via -r requirements.in
types-python-dateutil==2.8.19.20240106
    # via arrow
typing-extensions==4.10.0
    # via typing-inspect
typing-inspect==0.9.0
    # via marshmallow-dataclass
tzlocal==5.2
    # via dateparser
urllib3==2.0.7
    # via
    #   botocore
    #   requests
websockets==12.0
    # via -r requirements.in
wheel==0.42.0
    # via pip-tools
yarl==1.9.4
    # via
    #   aiohttp
    #   gql

# The following packages are considered to be unsafe in a requirements file:
# pip
# setuptools
